backend/cocktailgram/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.2
python-dotenv==1.0.1
django-filter==23.4
Pillow==10.1.0
drf-extra-fields==3.6.1
reportlab==4.0.9
gunicorn==20.1.0
django-cors-headers==3.13.0
psycopg2-binary==2.9.3 